+++ views/pool-create/select-coins copy/input/header-info.tsx
