+++ views/pool-create/select-coins copy/input/index.tsx
