+++ views/pool-create/select-coins copy/input/input.types.ts
