+++ views/pool-create/select-coins copy/input/select-token.tsx
